Revert "Adding more unlock keychain requests"
This reverts commit 6c6b18c160c4b6bff6f60e5b1b78df2041843e3c.
