+++ src/main/java/org/hardisonbrewing/maven/cxx/xcode/CompileMojo.java
@@ -36,11 +36,6 @@ public final class CompileMojo extends JoJoMojoImpl {
     /**
      * @parameter
      */
-    public Keychain keychain;
-
-    /**
-     * @parameter
-     */
     public String scheme;
 
     @Override
@@ -57,8 +52,6 @@ public final class CompileMojo extends JoJoMojoImpl {
     }
 
     private void execute( String target ) {
-
-        unlockKeychain( keychain );
 
         List<String> cmd = new LinkedList<String>();
         cmd.add( "xcodebuild" );
@@ -121,11 +114,5 @@ public final class CompileMojo extends JoJoMojoImpl {
         catch (CommandLineException e) {
             throw new IllegalStateException( e );
         }
-    }
-
-    private void unlockKeychain( Keychain keychain ) {
-
-        List<String> cmd = KeychainHelper.unlockKeychainCommand( keychain );
-        execute( cmd );
     }
 }
