+++ src/main/java/org/hardisonbrewing/maven/cxx/xcode/GenerateIpaMojo.java
@@ -66,8 +66,6 @@ public final class GenerateIpaMojo extends JoJoMojoImpl {
 
     private void generateIpaFile( String target ) {
 
-        unlockKeychain( keychain );
-
         List<String> cmd = new LinkedList<String>();
 
         cmd.add( "xcrun" );
@@ -150,11 +148,5 @@ public final class GenerateIpaMojo extends JoJoMojoImpl {
         stringBuffer.append( "." );
         stringBuffer.append( XCodeService.IPA_EXTENSION );
         return stringBuffer.toString();
-    }
-
-    private void unlockKeychain( Keychain keychain ) {
-
-        List<String> cmd = KeychainHelper.unlockKeychainCommand( keychain );
-        execute( cmd );
     }
 }
